Move where_block_number_in_period to Explorer.Chain.Block.Reader.General
