+++ apps/explorer/lib/explorer/chain.ex
@@ -66,6 +66,8 @@ defmodule Explorer.Chain do
     Withdrawal
   }
 
+  alias Explorer.Chain.Block.Reader.General, as: BlockReaderGeneral
+
   alias Explorer.Chain.Cache.{
     BlockNumber,
     Blocks,
@@ -207,7 +209,7 @@ defmodule Explorer.Chain do
             InternalTransaction
             |> InternalTransaction.where_nonpending_block()
             |> InternalTransaction.where_address_fields_match(hash, :to_address_hash)
-            |> where_block_number_in_period(from_block, to_block)
+            |> BlockReaderGeneral.where_block_number_in_period(from_block, to_block)
             |> common_where_limit_order(paging_options)
             |> wrapped_union_subquery()
 
@@ -215,7 +217,7 @@ defmodule Explorer.Chain do
             InternalTransaction
             |> InternalTransaction.where_nonpending_block()
             |> InternalTransaction.where_address_fields_match(hash, :from_address_hash)
-            |> where_block_number_in_period(from_block, to_block)
+            |> BlockReaderGeneral.where_block_number_in_period(from_block, to_block)
             |> common_where_limit_order(paging_options)
             |> wrapped_union_subquery()
 
@@ -223,7 +225,7 @@ defmodule Explorer.Chain do
             InternalTransaction
             |> InternalTransaction.where_nonpending_block()
             |> InternalTransaction.where_address_fields_match(hash, :created_contract_address_hash)
-            |> where_block_number_in_period(from_block, to_block)
+            |> BlockReaderGeneral.where_block_number_in_period(from_block, to_block)
             |> common_where_limit_order(paging_options)
             |> wrapped_union_subquery()
 
@@ -239,7 +241,7 @@ defmodule Explorer.Chain do
           InternalTransaction
           |> InternalTransaction.where_nonpending_block()
           |> InternalTransaction.where_address_fields_match(hash, direction)
-          |> where_block_number_in_period(from_block, to_block)
+          |> BlockReaderGeneral.where_block_number_in_period(from_block, to_block)
           |> common_where_limit_order(paging_options)
           |> preload(:block)
           |> join_associations(necessity_by_association)
@@ -402,7 +404,7 @@ defmodule Explorer.Chain do
         preloaded_query
         |> page_logs(paging_options)
         |> filter_topic(Keyword.get(options, :topic))
-        |> where_block_number_in_period(from_block, to_block)
+        |> BlockReaderGeneral.where_block_number_in_period(from_block, to_block)
         |> join_associations(necessity_by_association)
         |> select_repo(options).all()
         |> Enum.take(paging_options.page_size)
@@ -416,68 +418,6 @@ defmodule Explorer.Chain do
       where:
         log.first_topic == ^topic or log.second_topic == ^topic or log.third_topic == ^topic or
           log.fourth_topic == ^topic
-    )
-  end
-
-  @doc """
-  Filters the `base_query` to include only the records where the `block_number` falls within the specified period.
-
-  ## Parameters
-
-    - `base_query`: The initial query to be filtered.
-    - `from_block`: The starting block number of the period. Can be `nil`.
-    - `to_block`: The ending block number of the period. Can be `nil`.
-
-  ## Returns
-
-    - A query filtered by the specified block number period.
-
-  ## Examples
-
-    - When `from_block` is `nil` and `to_block` is not `nil`:
-      ```elixir
-      where_block_number_in_period(query, nil, 100)
-      # Filters the query to include records with block_number <= 100
-      ```
-
-    - When `from_block` is not `nil` and `to_block` is `nil`:
-      ```elixir
-      where_block_number_in_period(query, 50, nil)
-      # Filters the query to include records with block_number >= 50
-      ```
-
-    - When both `from_block` and `to_block` are `nil`:
-      ```elixir
-      where_block_number_in_period(query, nil, nil)
-      # Returns the base query without any filtering
-      ```
-
-    - When both `from_block` and `to_block` are not `nil`:
-      ```elixir
-      where_block_number_in_period(query, 50, 100)
-      # Filters the query to include records with block_number between 50 and 100 (inclusive)
-      ```
-  """
-  @spec where_block_number_in_period(Ecto.Query.t(), non_neg_integer() | nil, non_neg_integer() | nil) :: Ecto.Query.t()
-  def where_block_number_in_period(base_query, from_block, to_block) when is_nil(from_block) and not is_nil(to_block) do
-    from(q in base_query,
-      where: q.block_number <= ^to_block
-    )
-  end
-
-  def where_block_number_in_period(base_query, from_block, to_block) when not is_nil(from_block) and is_nil(to_block) do
-    from(q in base_query,
-      where: q.block_number >= ^from_block
-    )
-  end
-
-  def where_block_number_in_period(base_query, from_block, to_block) when is_nil(from_block) and is_nil(to_block) do
-    base_query
-  end
-
-  def where_block_number_in_period(base_query, from_block, to_block) do
-    from(q in base_query,
-      where: q.block_number >= ^from_block and q.block_number <= ^to_block
     )
   end
