+++ apps/explorer/lib/explorer/chain/block/reader/general.ex
@@ -190,4 +190,66 @@ defmodule Explorer.Chain.Block.Reader.General do
         end
     end
   end
+
+  @doc """
+  Filters the `base_query` to include only the records where the `block_number` falls within the specified period.
+
+  ## Parameters
+
+    - `base_query`: The initial query to be filtered.
+    - `from_block`: The starting block number of the period. Can be `nil`.
+    - `to_block`: The ending block number of the period. Can be `nil`.
+
+  ## Returns
+
+    - A query filtered by the specified block number period.
+
+  ## Examples
+
+    - When `from_block` is `nil` and `to_block` is not `nil`:
+      ```elixir
+      where_block_number_in_period(query, nil, 100)
+      # Filters the query to include records with block_number <= 100
+      ```
+
+    - When `from_block` is not `nil` and `to_block` is `nil`:
+      ```elixir
+      where_block_number_in_period(query, 50, nil)
+      # Filters the query to include records with block_number >= 50
+      ```
+
+    - When both `from_block` and `to_block` are `nil`:
+      ```elixir
+      where_block_number_in_period(query, nil, nil)
+      # Returns the base query without any filtering
+      ```
+
+    - When both `from_block` and `to_block` are not `nil`:
+      ```elixir
+      where_block_number_in_period(query, 50, 100)
+      # Filters the query to include records with block_number between 50 and 100 (inclusive)
+      ```
+  """
+  @spec where_block_number_in_period(Ecto.Query.t(), non_neg_integer() | nil, non_neg_integer() | nil) :: Ecto.Query.t()
+  def where_block_number_in_period(base_query, from_block, to_block) when is_nil(from_block) and not is_nil(to_block) do
+    from(q in base_query,
+      where: q.block_number <= ^to_block
+    )
+  end
+
+  def where_block_number_in_period(base_query, from_block, to_block) when not is_nil(from_block) and is_nil(to_block) do
+    from(q in base_query,
+      where: q.block_number >= ^from_block
+    )
+  end
+
+  def where_block_number_in_period(base_query, from_block, to_block) when is_nil(from_block) and is_nil(to_block) do
+    base_query
+  end
+
+  def where_block_number_in_period(base_query, from_block, to_block) do
+    from(q in base_query,
+      where: q.block_number >= ^from_block and q.block_number <= ^to_block
+    )
+  end
 end
